Create SessionAwareFragmentActivity to manage login screen redirection
